requirements.txt
appdirs
asgiref
attrs
backcall
black
certifi==2020.12.5
chardet==4.0.0
click
decorator
dj-database-url
Django
djangorestframework
djangorestframework-api-key==2.0.0
idna==2.10
iniconfig
ipdb
ipython
ipython-genutils
jedi
mypy-extensions
packaging
parso
pathspec
pexpect
pickleshare
pluggy
prompt-toolkit 
psycopg2-binary 
ptyprocess
py
Pygments
pyparsing 
pytest
pytest-django
python-decouple
pytz
regex
requests==2.25.1
sqlparse
toml
traitlets
typed-ast
typing-extensions 
urllib3==1.26.3
wcwidth
